Move MethodMatcher out of tracemethods package
