+++ apm-agent-core/src/main/java/co/elastic/apm/agent/configuration/CoreConfiguration.java
@@ -27,8 +27,8 @@ import co.elastic.apm.agent.configuration.validation.RegexValidator;
 import co.elastic.apm.agent.impl.transaction.Span;
 import co.elastic.apm.agent.matcher.WildcardMatcher;
 import co.elastic.apm.agent.matcher.WildcardMatcherValueConverter;
-import co.elastic.apm.agent.tracemethods.MethodMatcher;
-import co.elastic.apm.agent.tracemethods.configuration.MethodMatcherValueConverter;
+import co.elastic.apm.agent.matcher.MethodMatcher;
+import co.elastic.apm.agent.matcher.MethodMatcherValueConverter;
 import org.stagemonitor.configuration.ConfigurationOption;
 import org.stagemonitor.configuration.ConfigurationOptionProvider;
 import org.stagemonitor.configuration.converter.MapValueConverter;
