+++ apm-agent-core/src/main/java/co/elastic/apm/agent/matcher/MethodMatcherValueConverter.java
@@ -16,16 +16,13 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.apm.agent.tracemethods.configuration;
+package co.elastic.apm.agent.matcher;
 
 import co.elastic.apm.agent.configuration.converter.ListValueConverter;
-import co.elastic.apm.agent.sdk.state.GlobalState;
-import co.elastic.apm.agent.tracemethods.MethodMatcher;
 import org.stagemonitor.configuration.converter.ValueConverter;
 
 import static co.elastic.apm.agent.configuration.converter.ListValueConverter.COMMA_OUT_OF_BRACKETS;
 
-@GlobalState
 public enum MethodMatcherValueConverter implements ValueConverter<MethodMatcher> {
     INSTANCE;
