+++ apm-agent-core/src/test/java/co/elastic/apm/agent/matcher/MethodMatcherInstrumentationTest.java
@@ -16,12 +16,14 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.apm.agent.tracemethods;
+package co.elastic.apm.agent.matcher;
 
 import co.elastic.apm.agent.MockReporter;
 import co.elastic.apm.agent.configuration.SpyConfiguration;
 import co.elastic.apm.agent.impl.ElasticApmTracer;
 import co.elastic.apm.agent.impl.ElasticApmTracerBuilder;
+import co.elastic.apm.agent.matcher.MethodMatcher;
+import co.elastic.apm.agent.tracemethods.TraceMethodInstrumentation;
 import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.description.type.TypeDescription;
 import org.junit.jupiter.api.BeforeAll;
