+++ apm-agent-core/src/test/java/co/elastic/apm/agent/matcher/MethodMatcherTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.apm.agent.tracemethods;
+package co.elastic.apm.agent.matcher;
 
 import org.junit.jupiter.api.Test;
